Move common stuff used only by connector and bootstrap to connector
