+++ bootstrap/sponge/src/main/java/org/geysermc/platform/sponge/command/GeyserSpongeCommandExecutor.java
@@ -27,7 +27,7 @@ package org.geysermc.platform.sponge.command;
 
 import lombok.AllArgsConstructor;
 
-import org.geysermc.common.ChatColor;
+import org.geysermc.connector.common.ChatColor;
 import org.geysermc.connector.GeyserConnector;
 import org.geysermc.connector.command.GeyserCommand;
 import org.spongepowered.api.command.CommandCallable;
