+++ bootstrap/velocity/src/main/java/org/geysermc/platform/velocity/GeyserVelocityMain.java
@@ -26,7 +26,7 @@
 
 package org.geysermc.platform.velocity;
 
-import org.geysermc.common.main.IGeyserMain;
+import org.geysermc.connector.common.main.IGeyserMain;
 
 public class GeyserVelocityMain extends IGeyserMain {
