--- common/src/main/java/org/geysermc/common/PlatformType.java
@@ -1,17 +1,0 @@
-package org.geysermc.common;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum PlatformType {
-
-    BUNGEECORD("BungeeCord"),
-    SPIGOT("Spigot"),
-    SPONGE("Sponge"),
-    STANDALONE("Standalone"),
-    VELOCITY("Velocity");
-
-    private String platformName;
-}
