+++ connector/src/main/java/org/geysermc/connector/command/defaults/ReloadCommand.java
@@ -25,8 +25,8 @@
 
 package org.geysermc.connector.command.defaults;
 
-import org.geysermc.common.ChatColor;
-import org.geysermc.common.PlatformType;
+import org.geysermc.connector.common.ChatColor;
+import org.geysermc.connector.common.PlatformType;
 import org.geysermc.connector.GeyserConnector;
 import org.geysermc.connector.command.CommandSender;
 import org.geysermc.connector.command.GeyserCommand;
