+++ connector/src/main/java/org/geysermc/connector/configuration/GeyserJacksonConfiguration.java
@@ -30,7 +30,7 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
-import org.geysermc.common.serializer.AsteriskSerializer;
+import org.geysermc.connector.common.serializer.AsteriskSerializer;
 
 import java.nio.file.Path;
 import java.util.Map;
